Modification of some logics to reduce similar code.
Next step is to move the get_habits() function from habit.py to db.py.

Then, getting "db" out of the method of Habit would be nice (with db in the definition of the Habit class and self.db). I had some errors this morning so I cancelled it and I will do it again at a later point.

Still some docstrings and comments missing.