Moved get_habits from habit.py to db.py.
Then, getting "db" out of the method of Habit would be nice (with db in the definition of the Habit class and self.db). I had some errors this morning so I cancelled it and I will do it again at a later point.

Still some docstrings and comments missing but a lot has been done.

I have also separated the list of habits as global variable and the database to query less often the database and already have data locally.